refactor with builder pattern
