+++ src/lib.rs
@@ -1,8 +1,7 @@
-mod bit_array;
 mod bloom_filter;
 mod compressor;
 mod decoder;
 mod encoder;
 mod hash;
 
-pub use bloom_filter::{BloomFilter, CompressMode};
+pub use bloom_filter::{BloomFilter, BloomFilterBuilder, CompressMode};
